# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a ecommerce API",
    "version": "1.0.0",
    "title": "Ecommerce API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "okwukwe.ewurum@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "fcode-ecommerce.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/categories": {
      "post": {
        "tags": ["Categories"],
        "summary": "Create a new category",
        "description": "Create a new category",
        "operationId": "create",
        "consumes": ["application/json", "application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "The category object to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/newCategory"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "new category create",
            "schema": {
              "$ref": "#/components/schemas/categoryResponse"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/components/schemas/inputError"
            }
          },
          "401": {
            "description": "unauthorized",
            "schema": {
              "$ref": "#/components/schemas/authError"
            }
          }
        }
      },
      "get": {
        "tags": ["Categories"],
        "summary": "Get all categories",
        "description": "Get all categories",
        "operationId": "getAllCategories",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "An array of categories",
            "schema": {
              "$ref": "#/components/schemas/allCategoriesResponse"
            }
          },
          "404": {
            "description": "not found",
            "schema": {
              "$ref": "#/components/schemas/notFoundResponse"
            }
          }
        }
      }
    },
    "/categories/{category_id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get a category by ID",
        "description": "Get one category by the category ID",
        "operationId": "getCategoryById",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "An object of category",
            "schema": {
              "$ref": "#/components/schemas/categoryResponse"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/components/schemas/invalidIdResponse"
            }
          },
          "404": {
            "description": "not found",
            "schema": {
              "$ref": "#/components/schemas/notFoundResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "newCategory": {
      "type": "object",
      "required": ["name", "department_id"],
      "properties": {
        "name": {
          "type": "string",
          "example": "new category"
        },
        "description": {
          "type": "string",
          "example": "new category"
        },
        "department_id": {
          "type": "integer",
          "example": 1
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "inputError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "USR_03"
          },
          "message": {
            "type": "string",
            "example": "This field is not allowed to be empty"
          },
          "field": {
            "type": "string",
            "example": "example"
          },
          "status": {
            "type": "integer",
            "example": 400
          }
        }
      },
      "authError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "AUT_02"
          },
          "message": {
            "type": "string",
            "example": "The apikey is invalid."
          },
          "field": {
            "type": "string",
            "example": "API_KEY"
          },
          "status": {
            "type": "integer",
            "example": 401
          }
        }
      },
...
      "categoryResponse": {
        "type": "object",
        "properties": {
          "category_id": {
            "type": "interger",
            "example": 1
          },
          "department_id": {
            "type": "interger",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "new category"
          },
          "description": {
            "type": "string",
            "example": "new description"
          }
        }
      },
      "allCategoriesResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/categoryResponse"
        }
      },
      "notFoundResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "USR_03"
          },
          "message": {
            "type": "string",
            "example": "Does not exist example with ID 5"
          },
          "status": {
            "type": "integer",
            "example": 404
          }
        }
      },
      "invalidIdResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "USR_09"
          },
          "message": {
            "type": "string",
            "example": "The ID is not a number"
          },
          "status": {
            "type": "integer",
            "example": 400
          }
        }
      }
    }
  }
}